+++ packages/alpha/src/__tests__/api/simulatorRoutinesApi.spec.ts
@@ -5,13 +5,11 @@ import { setupLoggedInClient } from '../testUtils';
 import {
   fileExtensionTypes,
   stepFields,
-  unitsMap,
-  unitSystem,
   modelTypes,
-  boundaryConditions,
   routineRevisionConfiguration,
+  unitQuantities,
   routineRevisionScript,
-} from './mocks';
+} from './seed';
 
 const SHOULD_RUN_TESTS = process.env.RUN_SDK_SIMINT_TESTS == 'true';
 
@@ -20,11 +18,11 @@ const describeIf = SHOULD_RUN_TESTS ? describe : describe.skip;
 describeIf('simulator routines api', () => {
   const ts = Date.now();
   const simulatorExternalId = `test_sim_${ts}_c`;
-  const modelExternalId = `test_sim_model_${ts}`;
-  const modelRevisionExternalId = `test_sim_model_revision_${ts}`;
+  const modelExternalId = `test_sim_model_${ts}_2`;
+  const modelRevisionExternalId = `test_sim_model_revision_${ts}_2_1`;
   const routineExternalId = `test_sim_routine_${ts}`;
   const routineRevisionExternalId = `test_sim_routine_revision_${ts}`;
-  const simulatorIntegrationExternalId = `test_sim_integration_${ts}`;
+  const simulatorIntegrationExternalId = `test_sim_integration_${ts}_2`;
   const simulatorName = `TestSim - ${ts}`;
   const client: CogniteClientAlpha = setupLoggedInClient();
   let simulatorId: number;
@@ -35,15 +33,9 @@ describeIf('simulator routines api', () => {
         externalId: simulatorExternalId,
         name: simulatorName,
         fileExtensionTypes,
-        enabled: true,
         stepFields,
-        units: {
-          unitsMap,
-          unitSystem,
-        },
+        unitQuantities,
         modelTypes,
-        boundaryConditions,
-        isCalculationsEnabled: true,
       },
     ]);
     simulatorId = response[0].id;
@@ -57,12 +49,11 @@ describeIf('simulator routines api', () => {
         externalId: simulatorIntegrationExternalId,
         simulatorExternalId: simulatorExternalId,
         heartbeat: new Date(ts),
-        dataSetId: 4097666328084896,
+        dataSetId: 97552494921583,
         connectorVersion: '1.0.0',
         simulatorVersion: '1.0.0',
-        runApiEnabled: true,
-        licenseStatus: 'active',
-        connectorStatus: 'unknown',
+        licenseStatus: 'UNKNOWN',
+        connectorStatus: 'IDLE',
       },
     ]);
     expect(response.length).toBe(1);
@@ -76,10 +67,9 @@ describeIf('simulator routines api', () => {
         simulatorExternalId,
         name: 'Test Simulator Model',
         description: 'Test Simulator Model Desc',
-        dataSetId: 4097666328084896,
+        dataSetId: 97552494921583,
         labels: [{ externalId: 'air-quality-po-1' }],
-        type: 'string',
-        unitSystem: 'SI',
+        type: 'WaterWell',
       },
     ]);
     expect(res.length).toBe(1);
@@ -92,9 +82,7 @@ describeIf('simulator routines api', () => {
         externalId: modelRevisionExternalId,
         modelExternalId,
         description: 'test sim model revision description',
-        fileId: 3747718694331206,
-        boundaryConditions: [],
-        metadata: {},
+        fileId: 6396395402204465,
       },
     ]);
     expect(response.length).toBe(1);
@@ -139,6 +127,36 @@ describeIf('simulator routines api', () => {
     );
   });
 
+  test('create routine revision', async () => {
+    const revisionExternalId = `${routineRevisionExternalId}_2`;
+    const response = await client.simulators.createRoutineRevisions([
+      {
+        externalId: revisionExternalId,
+        routineExternalId,
+        configuration: routineRevisionConfiguration,
+        script: [],
+      },
+    ]);
+    expect(response.length).toBe(1);
+    expect(response[0].externalId).toBe(revisionExternalId);
+    expect(response[0].simulatorIntegrationExternalId).toBe(
+      simulatorIntegrationExternalId
+    );
+
+    const listRoutineRevisions = await client.simulators.listRoutineRevisions({
+      filter: {
+        routineExternalIds: [routineExternalId],
+        allVersions: false,
+      },
+    });
+
+    const revisionFilter = listRoutineRevisions.items.filter(
+      (item) => item.routineExternalId === routineExternalId
+    );
+
+    expect(revisionFilter.length).toBe(1);
+  });
+
   test('list routine revision', async () => {
     const listResponse = await client.simulators.listRoutineRevisions();
     expect(listResponse.items.length).toBeGreaterThan(0);
@@ -150,9 +168,9 @@ describeIf('simulator routines api', () => {
 
   test('list routine revision w filters', async () => {
     const listFilterResponse = await client.simulators.listRoutineRevisions({
-      filter: { routineExternalIds: [routineExternalId] },
+      filter: { routineExternalIds: [routineExternalId], allVersions: true },
     });
-    expect(listFilterResponse.items.length).toBe(1);
+    expect(listFilterResponse.items.length).toBe(2);
     const routineRevisionFound = listFilterResponse.items.find(
       (item) => item.externalId === routineRevisionExternalId
     );