+++ packages/beta/src/types.ts
@@ -18,6 +18,7 @@ import {
   StringDatapoints as StringDatapointsStable,
   Timestamp,
   SortOrder,
+  FileInfo,
 } from '@cognite/sdk';
 import {
   CogniteExternalId,
@@ -344,6 +345,16 @@ export interface Subscription {
   subscriberId: CogniteInternalId;
   subscriberExternalId?: CogniteExternalId;
   metadata?: Metadata;
+}
+
+export interface MultiPartFileUploadResponse extends FileInfo {
+  uploadUrls: string[];
+  uploadId: string;
+}
+
+export interface MultiPartFileChunkResponse {
+  partNumber: number;
+  status: number;
 }
 
 export interface DatapointsMultiQuery extends DatapointsMultiQueryBase {